// src/generated-types/media.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.0
//   protoc               v6.33.2
// source: media.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "media.v1";

/** FileKey represents the request message containing the file key. */
export interface FileKey {
  fileKey: string;
}

/** UploadAvatarRequest represents the request message for uploading an avatar. */
export interface UploadAvatarRequest {
  id: string;
  buffer: Uint8Array;
  fieldName: string;
  originalName: string;
  mimeType: string;
  size: number;
}

/** FileUrl represents the response message after uploading an avatar. */
export interface FileUrl {
  fileUrl: string;
}

/** declaration of Id message */
export interface Id {
  id: string;
}

/** Message representing a status response */
export interface StatusResponse {
  success: boolean;
  message: string;
}

export const MEDIA_V1_PACKAGE_NAME = "media.v1";

/** MediaService defines the gRPC service for media operations. */

export interface MediaServiceClient {
  /** rpc for getting image URL with file key */

  getImageUrl(request: FileKey): Observable<FileUrl>;

  /** rpc for uploading avatar file */

  uploadAvatar(request: UploadAvatarRequest): Observable<FileUrl>;

  /** rpc for deleting avatar file */

  deleteAvatar(request: FileKey): Observable<StatusResponse>;
}

/** MediaService defines the gRPC service for media operations. */

export interface MediaServiceController {
  /** rpc for getting image URL with file key */

  getImageUrl(request: FileKey): Promise<FileUrl> | Observable<FileUrl> | FileUrl;

  /** rpc for uploading avatar file */

  uploadAvatar(request: UploadAvatarRequest): Promise<FileUrl> | Observable<FileUrl> | FileUrl;

  /** rpc for deleting avatar file */

  deleteAvatar(request: FileKey): Promise<StatusResponse> | Observable<StatusResponse> | StatusResponse;
}

export function MediaServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["getImageUrl", "uploadAvatar", "deleteAvatar"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("MediaService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("MediaService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const MEDIA_SERVICE_NAME = "MediaService";